requirements.txt
Flask
gunicorn
flask-talisman
python-dotenv